+++ src/modules/submit/ReviewForm/ReviewForm.theme.ts
@@ -4,22 +4,7 @@ export type ReviewFormVariants = VariantProps<typeof reviewFormTheme>;
 
 export const reviewFormTheme = tv(
   {
-    slots: {
-      wrapper: [
-        "flex",
-        "w-[48rem]",
-        "flex-col",
-        "items-start",
-        "gap-6",
-        "rounded-2xl",
-        "bg-surface-base",
-        "px-6",
-        "py-8",
-      ],
-      divider: ["h-0", "w-full", "border-border-default"],
-      lower: ["flex", "flex-col", "items-start", "gap-8"],
-      textarea: ["w-full"],
-    },
+    base: ["inline-flex", "flex-col", "items-start", "gap-[3.125rem]"],
   },
   { responsiveVariants: true },
 );
