+++ src/modules/submit/ReviewForm/index.ts
@@ -1,3 +1,1 @@
 export * from "./ReviewForm";
-export * from "./ReviewForm.theme";
-export * from "./types";
